// src/app/core/models/models.ts
export interface pageSelection {
  skip: number;
  limit: number;
}